Move DomainSerializer into sub namespace
