+++ src/PrestaShopBundle/ApiPlatform/Processor/CommandProcessor.php
@@ -35,9 +35,9 @@ use PrestaShop\PrestaShop\Core\Context\ApiClientContext;
 use PrestaShop\PrestaShop\Core\Context\LanguageContext;
 use PrestaShop\PrestaShop\Core\Context\ShopContext;
 use PrestaShopBundle\ApiPlatform\ContextParametersTrait;
-use PrestaShopBundle\ApiPlatform\DomainSerializer;
 use PrestaShopBundle\ApiPlatform\Exception\CQRSCommandNotFoundException;
 use PrestaShopBundle\ApiPlatform\QueryResultSerializerTrait;
+use PrestaShopBundle\ApiPlatform\Serializer\DomainSerializer;
 use ReflectionException;
 use Symfony\Component\Serializer\Exception\ExceptionInterface;
