+++ src/PrestaShopBundle/ApiPlatform/Provider/QueryListProvider.php
@@ -39,10 +39,10 @@ use PrestaShop\PrestaShop\Core\Grid\Data\Factory\GridDataFactoryInterface;
 use PrestaShop\PrestaShop\Core\Search\Builder\FiltersBuilderInterface;
 use PrestaShop\PrestaShop\Core\Search\Filters;
 use PrestaShopBundle\ApiPlatform\ContextParametersTrait;
-use PrestaShopBundle\ApiPlatform\DomainSerializer;
 use PrestaShopBundle\ApiPlatform\Exception\GridDataFactoryNotFoundException;
 use PrestaShopBundle\ApiPlatform\Pagination\PaginationElements;
 use PrestaShopBundle\ApiPlatform\QueryResultSerializerTrait;
+use PrestaShopBundle\ApiPlatform\Serializer\DomainSerializer;
 use Psr\Container\ContainerInterface;
 use ReflectionException;
 use Symfony\Component\HttpFoundation\RequestStack;
